+++ src/app/components/settings-form/settings-form.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, Output, inject, signal } from '@angular/core';
+import { Component, inject, signal } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatExpansionModule } from '@angular/material/expansion';
@@ -13,7 +13,6 @@ import { PricePerMTokensPipe } from '../../pipes/price-per-m-tokens.pipe';
 import { UpperFirstLetterPipe } from '../../pipes/upper-first-letter.pipe';
 import { SettingsService } from '../../services/settings.service';
 import { Model } from '../../types/model.types';
-import { ModelProvider } from '../../../assets/config/models';
 import { TaskTypeId } from '../../../assets/config/prompts';
 
 @Component({
@@ -36,8 +35,6 @@ import { TaskTypeId } from '../../../assets/config/prompts';
 })
 export class SettingsFormComponent {
   settings = inject(SettingsService);
-
-  @Output() providerChanged = new EventEmitter<ModelProvider>();
 
   teiEncode = signal<boolean>(false);
   transcribeHeaders = signal<boolean>(true);
@@ -74,14 +71,7 @@ export class SettingsFormComponent {
   }
 
   setModel(model: Model): void {
-    const prevProvider = this.settings.selectedModel().provider;
-
     this.settings.updateSelectedModel(model);
-
-    const nextProvider = model.provider;
-    if (nextProvider !== prevProvider) {
-      this.providerChanged.emit(nextProvider);
-    }
   }
 
   setTaskType(type: TaskTypeId): void {